Cryptography/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
argon2-cffi==23.1.0
bcrypt==4.1.1
pytest==7.4.3
pytest-cov==4.1.0
werkzeug==3.0.1
python-dotenv==1.0.0
mailjet-rest==1.3.4

